feat(schedule): update after merge review

+++ src/main/java/fr/gouv/stopc/submission/code/server/business/service/ViewsService.java
@@ -114,14 +114,14 @@ public class ViewsService {
             ViewVo.CodeGenerationRequestBody codeGenerationRequestBody) throws SubmissionCodeServerException {
 
         @NotNull
-        final long codePerDay = codeGenerationRequestBody.getDailyAmount();
+        final Long codePerDay = codeGenerationRequestBody.getDailyAmount();
         @NotNull
         final OffsetDateTime from = codeGenerationRequestBody.getFrom();
         @NotNull
         OffsetDateTime to = codeGenerationRequestBody.getTo();
 
         Lot newLot = new Lot();
-        newLot.setNumberOfCodes(codePerDay);
+        newLot.setNumberOfCodes(codePerDay.intValue());
         newLot.setDateExecution(OffsetDateTime.now());
         this.fileExportService.zipExportAsync(
                 codePerDay,
