+++ src/main/java/fr/gouv/stopc/submission/code/server/business/service/schedule/GenerationRequest.java
@@ -17,7 +17,7 @@ public class GenerationRequest {
 
     private OffsetDateTime endDateTime;
 
-    private Long numberOfCodeToGenerate;
+    private Integer numberOfCodeToGenerate;
 
     private Long iterationBatchNumber;
 }
